api/requirements.txt
Django==2.0.4
psycopg2-binary==2.7.4
dj-database-url==0.5.0
prettyconf==1.2.3
factory-boy==2.10.0